Feature/create config map for server

+++ csharp/Common/Models/Influx/InfluxRecord.cs
@@ -2,7 +2,7 @@ namespace Common.Models.Influx;
 
 public record InfluxRecord{
     public DateTime Time { get; set; }
-    public double Watt { get; set; }
+    public float Watt { get; set; }
     public double Humidity { get; set; }
     public double Pressure { get; set; }
     public double Temperature { get; set; }
